feat: add auth on book route
